[add]HeaderにLinkがあるページのガワを作成
HeaderにLinkがあるHome,Profile,Searchページを作成しました。
