+++ src/app/globals.scss
@@ -80,6 +80,10 @@ header {
   width: 260px;
   height: 100vh;
   position: fixed;
+
+  @media (max-width: 1264px) {
+    width: 75px;
+  }
 }
 
 main {
@@ -87,6 +91,11 @@ main {
   margin-left: 260px;
   justify-content: center;
   display: flex;
+
+  @media (max-width: 1264px) {
+    width: calc(100% - 75px);
+    margin-left: 75px;
+  }
 }
 
 * {
