+++ src/app/home/page.tsx
@@ -1,0 +1,12 @@
+import { NextPage } from "next";
+
+const HomePage: NextPage = () => {
+  return (
+    <div>
+      <h1>ホームページ</h1>
+      <p>ここにホームページの内容が表示されます。</p>
+    </div>
+  );
+};
+
+export default HomePage;
