+++ src/app/profile/page.tsx
@@ -1,0 +1,12 @@
+import { NextPage } from "next";
+
+const ProfilePage: NextPage = () => {
+  return (
+    <div>
+      <h1>プロフィールページ</h1>
+      <p>ここにプロフィールページの内容が表示されます。</p>
+    </div>
+  );
+};
+
+export default ProfilePage;
