+++ src/app/swipe/result/page.module.scss
