+++ src/app/swipe/search/page.module.scss
