+++ src/components/base/Header/Header.module.scss
@@ -1,5 +1,4 @@
 .container {
-  width: 260px;
   display: flex;
   flex-flow: column;
   justify-content: space-between;
@@ -29,6 +28,10 @@
       span {
         margin-left: 15px;
         font-weight: bold;
+
+        @media (max-width: 1264px) {
+          display: none;
+        }
       }
     }
   }
@@ -52,13 +55,51 @@
       margin: 5px 0;
 
       .image {
-        margin: 5px;
         border-radius: 20px;
+        width: 40px;
+        height: 40px;
       }
 
       span {
-        margin-left: 15px;
+        margin-left: 20px;
         font-weight: bold;
+        font-size: 18px;
+
+        @media (max-width: 1264px) {
+          display: none;
+        }
+      }
+    }
+
+    .authContainer {
+      padding: 10px;
+      .authLink {
+        padding: 10px 50px;
+        @media (max-width: 1264px) {
+          display: none;
+        }
+      }
+      .hamburger {
+        display: none;
+        margin: 5px;
+        position: relative;
+
+        @media (max-width: 1264px) {
+          display: block;
+        }
+      }
+
+      .menu {
+        position: absolute;
+        top: -75%;
+        left: 150%;
+        background-color: #000000;
+        border: 1px solid #262626;
+        border-radius: 4px;
+        width: 200px;
+        padding: 20px;
+        z-index: 1000;
+        box-shadow: 0px 0px 10px 5px rgba(59, 59, 59, 0.3);
       }
     }
   }
